test: add unit tests for componentDidMount, init, getOffsetByValue and getMaxOffset methods
